+++ modules/nlu/src/backend/engine2/intent-vocab.test.ts
@@ -45,8 +45,8 @@ describe('Build intent vocab', () => {
   })
 
   test('With utterance tokens only', () => {
-    const u1 = new Utterance(u1Toks, genMockVectors(u1Toks))
-    const u2 = new Utterance(u2Toks, genMockVectors(u2Toks))
+    const u1 = new Utterance(u1Toks, genMockVectors(u1Toks), 'en')
+    const u2 = new Utterance(u2Toks, genMockVectors(u2Toks), 'en')
 
     const intVocab = buildIntentVocab([u1, u2], [])
     const allUtoks = [...u1Toks, ...u2Toks]
@@ -66,8 +66,8 @@ describe('Build intent vocab', () => {
   })
 
   test('With list entitites and Utterance tokens', () => {
-    const u1 = new Utterance(u1Toks, genMockVectors(u1Toks))
-    const u2 = new Utterance(u2Toks, genMockVectors(u2Toks))
+    const u1 = new Utterance(u1Toks, genMockVectors(u1Toks), 'en')
+    const u2 = new Utterance(u2Toks, genMockVectors(u2Toks), 'en')
 
     const intVocab = buildIntentVocab([u1, u2], LIST_ENTITIES)
     const allUtoks = [...u1Toks, ...u2Toks]
@@ -89,7 +89,7 @@ describe('Build intent vocab', () => {
   })
 
   test('Some tokens with tagged slots', () => {
-    const u1 = new Utterance(u1Toks, genMockVectors(u1Toks))
+    const u1 = new Utterance(u1Toks, genMockVectors(u1Toks), 'en')
     u1.tagSlot({ name: 'person' } as ExtractedSlot, 6, 16) // slot is: "my friend"
 
     const intVocab = buildIntentVocab([u1], [])