+++ modules/nlu/src/backend/engine2/labeler2.test.ts
@@ -5,7 +5,7 @@ import Utterance from './utterance'
 describe('CRF labels for utterance', () => {
   test('without slots', () => {
     const toks = 'My mame is Heisenberg and I am the danger'.split(/(\s)/g)
-    const utterance = new Utterance(toks, new Array(toks.length).fill([0]))
+    const utterance = new Utterance(toks, new Array(toks.length).fill([0]), 'en')
 
     const labels = labelizeUtterance(utterance)
 
@@ -17,7 +17,7 @@ describe('CRF labels for utterance', () => {
     const toks = 'Careful my friend, Alex W. is one of us'.split(/(\s)/g)
     //            012345678901234567890123456789012345678
     //            ________---------__-------___________--
-    const utterance = new Utterance(toks, new Array(toks.length).fill([0]))
+    const utterance = new Utterance(toks, new Array(toks.length).fill([0]), 'en')
     utterance.tagSlot({ name: 'listener', source: 'my friend' } as ExtractedSlot, 8, 18) // 18 because we want to include the ',' (for testing purposes) since we're not tokenizing wisely
     utterance.tagEntity({ value: 'my friend', type: 'friend' } as ExtractedEntity, 8, 18) // 18 because we want to include the ',' (for testing purposes) since we're not tokenizing wisely
     utterance.tagSlot({ name: 'person', source: 'Alex W.' } as ExtractedSlot, 19, 26)